tests for select - separate it because of element-ui bug
